Move Forge-related code to src/protocol/forge.rs
